Add TodoControl to manage Todo items with Folder support
The TodoControl class with matching UI is added to provide users with the ability to manage their tasks. It includes the functionality to create new folders to categorize the tasks, select a folder to add tasks to, and update configurations related to task management. Also, an icon picker dialog is included to customize the folder icons.

The change also introduces necessary data models for Todo control and support for managing these models. TodoData is relocated under the Todos namespace and updated to include folders and settings. TodoSettings and FolderItem models are introduced to handle specific configurations and folder data respectively.

Also, this diff changes the app's navigation to navigate to the new TodoControl. Further, this diff updates ConfigData to import the Todos namespace. The naming of the TodoItem model file is modified. In CubeMangerWindow.xaml, the height and width are set to auto and minimum dimensions are lowered.

These changes were needed to improve task management capabilities in the CubeManager application, making it more user-friendly and functionally rich.