Add Autorenew options to Todo items
Implemented Autorenewal functionality for Todo tasks to allow them to be set to renew daily, weekly, monthly or yearly. Included an option to disable renewal. Created this feature by adding a new button in the XAML file for the UI and the corresponding business logic in the CS file. This feature improves task management by efficiently handling recurrent tasks.