Add default property and context menu to folders
This update primarily adds a 'isDefault' property to FolderItem class. With this, the TodoControl can now determine whether a folder is default or not. This information is used to restrict user from deleting default folders.

To implement this, a context menu was added to FolderList items. On right-click, if the FolderItem is not default, a delete option is shown. Clicking it removes the folder both from TodoData.Folders and FolderList items.

Lastly, changes were made to ensure that FolderComboBox and FolderList have consistent states on selection change.