Add daily quote feature to UI
A new feature has been added to the UI where a quote from ZenQuotes API is fetched and displayed on the UI daily. This was implemented by creating new classes under the ZenQuotes namespace. The QuoteData class was added to the ConfigData class manifest for data management. A daily quote is fetched in the FetchQuote class by making a GET request to the ZenQuotes API. A 'day' check is implemented to ensure the quote is updated daily only. The quote is then saved via the SaveQuote method in the DayHelper. The application and package icons were also edited.