"Add SplashScreen with quote functionality"
Implemented a new SplashScreen to enhance program's startup experience. Quote retrieval was moved from CubeManagerWindow to this new SplashScreen, which now displays a quote whilst loading. Added a check to confirm the quote API is responsive before attempting to retrieve a quote. A loading ProgressBar was also implemented to provide the user with a visualization of loading progress. The initial window now leads to this SplashScreen instead of straight to CubeManagerWindow.