Simplify SplashScreen UI and implement new Login/Register pages
Reduced the complexity of CubeManager/SplashScreen.xaml for an improved user experience. Removed several elements in the splash screen and kept only the loading screen for a sleeker, focused design.

Additional files were added to the CubeManager/LoginRegister directory for the implementation of the new Login and Register windows. This is an essential step in enabling user authentication in the app.

Also updated CubeManager.csproj to include SkiaSharp libraries that'll be used for future UI elements.

Lastly, moved app startup sequence from CubeManagerWindow to LoginWindow in SplashScreen.xaml.cs, making the Login window the first page after the SplashScreen.