"Refactor LoginWindow rendering for performance improvement"
This commit optimizes the rendering performance of the LoginWindow. This was achieved by replacing the CompositionTarget.Rendering event with a DispatcherTimer at 60FPS, which gives the same rendering rate but with more control.

The _sweepShader was also updated to be created once, instead of in every frame, with a rotated copy of it used in each frame. This reduces unnecessary object creation.

Adding a distance check in the MouseMove event also reduces unnecessary visual invalidations, further improving performance.