Update sound settings UI and controls
Added sound settings controls to the settings window and updated the visibility logic. The changes provide users with an interface to control individual sound settings, and improve app usability by only displaying these controls when the 'Sound' or 'Dopamine' switches are checked. It includes options to change sound, play sound and reset sound for a number of events within the application. The visibility logic ensures that the controls are hidden from view when they are not relevant, preventing clutter and confusion.