Refactor Logger to use Serilog library
Logger.cs was refactored to use the Serilog library instead of custom logging mechanisms. This provides a more standardized and robust logging framework. The custom severity levels and their associated colors have been removed and standard Serilog methods (Information, Warning, Debug, Error, and Fatal) were used in place. The Serilog logger is configured to write logs to the console and to a rolling file in the application data folder. Old custom logs were also deleted as part of this update.