Remove StatusRequests enum and update several components
The StatusRequests enum has been removed, causing updates in classes that were using this type. CubeMessageBox's display settings in taskbar have been modified to true, and BaseResponse class has been updated to exclude Status. Additionally, the URL in the APICalls class was updated to include versioning and the success message text in registration is now hard-coded.